pyproject.toml
[project]
name = "bastion"
version = "0.1.0"
description = "Bastion ingestion harness — Block A scrapers and provenance writers"
requires-python = ">=3.11"
dependencies = [
    "psycopg[binary]>=3.2",     # Postgres driver (psycopg3, async-capable)
    "httpx>=0.27",              # HTTP client (sync + async)
    "beautifulsoup4>=4.12",     # HTML parsing
    "lxml>=5.0",                # fast parser backend
    "playwright>=1.45",         # JS-rendered pages
    "scrapy>=2.11",             # bulk crawls
    "tenacity>=8.2",            # retry with backoff
    "structlog>=24.1",          # structured logging
    "python-dotenv>=1.0",
    "click>=8.1",               # CLI
    "pyyaml>=6.0",
]

[project.scripts]
bastion = "bastion.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bastion"]
